+++ packages/plugin-image/web/src/imageStudio/image-studio-opener-loader.js
@@ -15,9 +15,34 @@ const IMAGE_STUDIO_OPENER_SRC =
 let mediaImageStudioPackageLoader;
 function getImageStudioPackage() {
   if (!mediaImageStudioPackageLoader) {
+    disableRequireJS(); //if requireJS is present the script doesn't load MediaImageStudio onto the window
     mediaImageStudioPackageLoader = loadScript(IMAGE_STUDIO_OPENER_SRC);
   }
-  return mediaImageStudioPackageLoader.then(() => window.MediaImageStudio);
+  return mediaImageStudioPackageLoader.then(() => {
+    enableRequireJS();
+    return window.MediaImageStudio;
+  });
+}
+
+function disableRequireJS() {
+  if (window.requirejs && !window.requirejsBackup) {
+    window.requirejsBackup = {
+      define: window.define,
+      require: window.require,
+      requirejs: window.requirejs,
+    };
+
+    window.define = undefined;
+    window.require = undefined;
+    window.requirejs = undefined;
+  }
+}
+
+function enableRequireJS() {
+  if (window.requirejsBackup) {
+    Object.assign(window, window.requirejsBackup);
+    window.requirejsBackup = undefined;
+  }
 }
 
 export { getImageStudioPackage };